+++ optimum/bettertransformer/models/__init__.py
@@ -49,24 +49,3 @@ def is_module_fast(module_name):
         return False
     else:
         return FAST_LAYERS_MAPPING_DICT[module_name]
-
-
-def convert_to_hf_classes(mapping_dict):
-    import transformers
-
-    hf_names_dict = {}
-    for fast_layer_key in mapping_dict.keys():
-        # For enc-decoder models the prefix is different
-        if "EncoderLayer" in fast_layer_key:
-            prefix = fast_layer_key[:-12]
-        else:
-            prefix = fast_layer_key[:-5]
-
-        # some `PreTrainedModel` models are not registerd in the auto mapping
-        if hasattr(transformers, prefix + "PreTrainedModel"):
-            hf_class = getattr(transformers, prefix + "PreTrainedModel")
-        else:
-            hf_class = getattr(transformers, prefix + "Model")
-
-        hf_names_dict[fast_layer_key] = hf_class
-    return hf_names_dict
