+++ tests/bettertransformers/test_bettertransformers.py
@@ -19,8 +19,9 @@ import torch
 import transformers
 from transformers import AutoModel, AutoModelForMaskedLM, AutoTokenizer, pipeline
 
-from optimum.bettertransformer import FAST_LAYERS_MAPPING_DICT, BetterTransformer, convert_to_hf_classes
+from optimum.bettertransformer import FAST_LAYERS_MAPPING_DICT, BetterTransformer
 from optimum.utils.testing_utils import (
+    convert_to_hf_classes,
     is_accelerate_available,
     is_torch_greater_than_113,
     require_accelerate,
